BED-5532 - PostgreSQL Batch Update (#1199)
* chore: refactor, cleanup and dry work

* chore+fix: further refactor and cleanup; BED-5532 batch operation behavior change to move away from transaction thrashing in pg
