+++ packages/go/cypher/models/pgsql/translate/function.go
@@ -213,7 +213,7 @@ func (s *Translator) translateFunction(typedExpression *cypher.FunctionInvocatio
 		} else {
 			switch typedArgument := unwrapParenthetical(argument).(type) {
 			case pgsql.Identifier:
-				if binding, bound := s.query.Scope.Lookup(typedArgument); !bound {
+				if binding, bound := s.scope.Lookup(typedArgument); !bound {
 					s.SetError(fmt.Errorf("binding not found for collect function argument %s", typedExpression.Name))
 				} else if bindingArrayType, err := binding.DataType.ToArrayType(); err != nil {
 					s.SetError(err)
@@ -239,4 +239,126 @@ func (s *Translator) translateFunction(typedExpression *cypher.FunctionInvocatio
 	default:
 		s.SetErrorf("unknown cypher function: %s", typedExpression.Name)
 	}
+}
+
+func (s *Translator) translateDateTimeFunctionCall(cypherFunc *cypher.FunctionInvocation, dataType pgsql.DataType) error {
+	// Ensure the local date time function uses the default precision
+	const defaultTimestampPrecision = 6
+
+	var functionIdentifier pgsql.Identifier
+
+	switch dataType {
+	case pgsql.Date:
+		functionIdentifier = pgsql.FunctionCurrentDate
+
+	case pgsql.TimeWithoutTimeZone:
+		functionIdentifier = pgsql.FunctionLocalTime
+
+	case pgsql.TimeWithTimeZone:
+		functionIdentifier = pgsql.FunctionCurrentTime
+
+	case pgsql.TimestampWithoutTimeZone:
+		functionIdentifier = pgsql.FunctionLocalTimestamp
+
+	case pgsql.TimestampWithTimeZone:
+		functionIdentifier = pgsql.FunctionNow
+
+	default:
+		return fmt.Errorf("unable to convert date function with data type: %s", dataType)
+	}
+
+	// Apply defaults for this function
+	if !cypherFunc.HasArguments() {
+		switch functionIdentifier {
+		case pgsql.FunctionCurrentDate:
+			s.treeTranslator.Push(pgsql.FunctionCall{
+				Function: functionIdentifier,
+				Bare:     true,
+				CastType: dataType,
+			})
+
+		case pgsql.FunctionNow:
+			s.treeTranslator.Push(pgsql.FunctionCall{
+				Function: functionIdentifier,
+				Bare:     false,
+				CastType: dataType,
+			})
+
+		default:
+			if precisionLiteral, err := pgsql.AsLiteral(defaultTimestampPrecision); err != nil {
+				return err
+			} else {
+				s.treeTranslator.Push(pgsql.FunctionCall{
+					Function: functionIdentifier,
+					Parameters: []pgsql.Expression{
+						precisionLiteral,
+					},
+					CastType: dataType,
+				})
+			}
+		}
+	} else if cypherFunc.NumArguments() > 1 {
+		return fmt.Errorf("expected only one text argument for cypher function: %s", cypherFunc.Name)
+	} else if specArgument, err := s.treeTranslator.Pop(); err != nil {
+		return err
+	} else {
+		s.treeTranslator.Push(pgsql.NewTypeCast(specArgument, dataType))
+	}
+
+	return nil
+}
+
+func (s *Translator) translateCoalesceFunction(functionInvocation *cypher.FunctionInvocation) error {
+	if numArgs := functionInvocation.NumArguments(); numArgs == 0 {
+		s.SetError(fmt.Errorf("expected at least one argument for cypher function: %s", functionInvocation.Name))
+	} else {
+		var (
+			arguments    = make([]pgsql.Expression, numArgs)
+			expectedType = pgsql.UnsetDataType
+		)
+
+		// This loop is used to pop off the coalesce function arguments in the intended order (since they're
+		// pushed onto the translator stack).
+		for idx := range functionInvocation.Arguments {
+			if argument, err := s.treeTranslator.Pop(); err != nil {
+				return err
+			} else {
+				arguments[numArgs-idx-1] = argument
+			}
+		}
+
+		// Find and validate types of the arguments
+		for _, argument := range arguments {
+			// Properties have no type information and should be skipped
+			if argumentType, err := InferExpressionType(argument); err != nil {
+				return err
+			} else if argumentType.IsKnown() {
+				// If the expected type isn't known yet then assign the known inferred type to it
+				if !expectedType.IsKnown() {
+					expectedType = argumentType
+				} else if expectedType != argumentType {
+					// All other inferrable argument types must match the first inferred type encountered
+					return fmt.Errorf("types in coalesce function must match %s but got %s", expectedType, argumentType)
+				}
+			}
+		}
+
+		if expectedType.IsKnown() {
+			// Rewrite any property lookup operators now that we have some type information
+			for idx, argument := range arguments {
+				if propertyLookup, isPropertyLookup := asPropertyLookup(argument); isPropertyLookup {
+					arguments[idx] = rewritePropertyLookupOperator(propertyLookup, expectedType)
+				}
+			}
+		}
+
+		// Translate the function call to the expected SQL form
+		s.treeTranslator.Push(pgsql.FunctionCall{
+			Function:   pgsql.FunctionCoalesce,
+			Parameters: arguments,
+			CastType:   expectedType,
+		})
+	}
+
+	return nil
 }
