+++ packages/go/cypher/models/pgsql/translate/projection.go
@@ -19,6 +19,8 @@ package translate
 import (
 	"fmt"
 
+	"github.com/specterops/bloodhound/cypher/models/cypher"
+
 	"github.com/specterops/bloodhound/cypher/models"
 	"github.com/specterops/bloodhound/cypher/models/pgsql"
 )
@@ -328,4 +330,178 @@ func buildProjection(alias pgsql.Identifier, projected *BoundIdentifier, scope *
 			Alias:      pgsql.AsOptionalIdentifier(alias),
 		},
 	}, nil
+}
+
+func (s *Translator) buildInlineProjection(part *QueryPart) (pgsql.Select, error) {
+	sqlSelect := pgsql.Select{
+		Where: part.projections.Constraints,
+	}
+
+	// If there's a projection frame set, some additional negotiation is required to identify which frame the
+	// from-statement should be written to. Some of this would be better figured out during the translation
+	// of the projection where query scope and other components are not yet fully translated.
+	if part.projections.Frame != nil {
+		// Look up to see if there are CTE expressions registered. If there are then it is likely
+		// there was a projection between this CTE and the previous multipart query part
+		hasCTEs := part.Model.CommonTableExpressions != nil && len(part.Model.CommonTableExpressions.Expressions) > 0
+
+		if part.Frame.Previous == nil || hasCTEs {
+			sqlSelect.From = []pgsql.FromClause{{
+				Source: part.projections.Frame.Binding.Identifier,
+			}}
+		} else {
+			sqlSelect.From = []pgsql.FromClause{{
+				Source: part.Frame.Previous.Binding.Identifier,
+			}}
+		}
+	}
+
+	for _, projection := range part.projections.Items {
+		builtProjection := projection.SelectItem
+
+		if projection.Alias.Set {
+			builtProjection = &pgsql.AliasedExpression{
+				Expression: builtProjection,
+				Alias:      projection.Alias,
+			}
+		}
+
+		sqlSelect.Projection = append(sqlSelect.Projection, builtProjection)
+	}
+
+	if len(part.projections.GroupBy) > 0 {
+		for _, groupBy := range part.projections.GroupBy {
+			sqlSelect.GroupBy = append(sqlSelect.GroupBy, groupBy)
+		}
+	}
+
+	return sqlSelect, nil
+}
+
+func (s *Translator) buildTailProjection() error {
+	var (
+		currentPart           = s.query.CurrentPart()
+		currentFrame          = s.scope.CurrentFrame()
+		singlePartQuerySelect = pgsql.Select{}
+	)
+
+	singlePartQuerySelect.From = []pgsql.FromClause{{
+		Source: pgsql.TableReference{
+			Name: pgsql.CompoundIdentifier{currentFrame.Binding.Identifier},
+		},
+	}}
+
+	if projectionConstraint, err := s.treeTranslator.ConsumeAll(); err != nil {
+		return err
+	} else if projection, err := buildExternalProjection(s.scope, currentPart.projections.Items); err != nil {
+		return err
+	} else if err := RewriteFrameBindings(s.scope, projectionConstraint.Expression); err != nil {
+		return err
+	} else {
+		singlePartQuerySelect.Projection = projection
+		singlePartQuerySelect.Where = projectionConstraint.Expression
+	}
+
+	currentPart.Model.Body = singlePartQuerySelect
+
+	if currentPart.Skip.Set {
+		currentPart.Model.Offset = currentPart.Skip
+	}
+
+	if currentPart.Limit.Set {
+		currentPart.Model.Limit = currentPart.Limit
+	}
+
+	if len(currentPart.OrderBy) > 0 {
+		currentPart.Model.OrderBy = currentPart.OrderBy
+	}
+
+	return nil
+}
+
+func (s *Translator) translateProjectionItem(scope *Scope, projectionItem *cypher.ProjectionItem) error {
+	if alias, hasAlias, err := extractIdentifierFromCypherExpression(projectionItem); err != nil {
+		return err
+	} else if nextExpression, err := s.treeTranslator.Pop(); err != nil {
+		return err
+	} else if selectItem, isProjection := nextExpression.(pgsql.SelectItem); !isProjection {
+		s.SetErrorf("invalid type for select item: %T", nextExpression)
+	} else {
+		if identifiers, err := ExtractSyntaxNodeReferences(selectItem); err != nil {
+			return err
+		} else if identifiers.Len() > 0 {
+			// Identifier lookups will require a scope reference
+			s.query.CurrentPart().projections.Frame = s.scope.CurrentFrame()
+		}
+
+		switch typedSelectItem := unwrapParenthetical(selectItem).(type) {
+		case pgsql.Identifier:
+			// If this is an identifier then assume the identifier as the projection alias since the translator
+			// rewrites all identifiers
+			if !hasAlias {
+				if boundSelectItem, bound := scope.Lookup(typedSelectItem); !bound {
+					return fmt.Errorf("invalid identifier: %s", typedSelectItem)
+				} else {
+					s.query.CurrentPart().CurrentProjection().SetAlias(boundSelectItem.Aliased())
+				}
+			}
+
+		case *pgsql.BinaryExpression:
+			// Binary expressions are used when properties are returned from a result projection
+			// e.g. match (n) return n.prop
+			if propertyLookup, isPropertyLookup := asPropertyLookup(typedSelectItem); isPropertyLookup {
+				// Ensure that projections maintain the raw JSONB type of the field
+				propertyLookup.Operator = pgsql.OperatorJSONField
+			}
+
+		default:
+			if hasAlias {
+				if inferredType, err := InferExpressionType(typedSelectItem); err != nil {
+					return err
+				} else if _, isBound := s.scope.AliasedLookup(alias); !isBound {
+					if newBinding, err := s.scope.DefineNew(inferredType); err != nil {
+						return err
+					} else {
+						// This binding is its own alias
+						s.scope.Alias(alias, newBinding)
+					}
+				}
+			}
+		}
+
+		if hasAlias {
+			s.query.CurrentPart().CurrentProjection().SetAlias(alias)
+		}
+
+		s.query.CurrentPart().CurrentProjection().SelectItem = selectItem
+	}
+
+	return nil
+}
+
+func (s *Translator) prepareProjection(projection *cypher.Projection) error {
+	currentPart := s.query.CurrentPart()
+	currentPart.PrepareProjections(projection.Distinct)
+
+	if projection.Skip != nil {
+		if cypherLiteral, isLiteral := projection.Skip.Value.(*cypher.Literal); !isLiteral {
+			return fmt.Errorf("expected a literal skip value but received: %T", projection.Skip.Value)
+		} else if pgLiteral, err := pgsql.AsLiteral(cypherLiteral.Value); err != nil {
+			return err
+		} else {
+			currentPart.Skip = models.ValueOptional[pgsql.Expression](pgLiteral)
+		}
+	}
+
+	if projection.Limit != nil {
+		if cypherLiteral, isLiteral := projection.Limit.Value.(*cypher.Literal); !isLiteral {
+			return fmt.Errorf("expected a literal limit value but received: %T", projection.Limit.Value)
+		} else if pgLiteral, err := pgsql.AsLiteral(cypherLiteral.Value); err != nil {
+			return err
+		} else {
+			currentPart.Limit = models.ValueOptional[pgsql.Expression](pgLiteral)
+		}
+	}
+
+	return nil
 }
